pyproject.toml
[project]
name = "mal-data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-dtypes>=1.4.4",
    "duckdb>=1.5.2",
    "google-auth>=2.49.2",
    "google-cloud-bigquery>=3.41.0",
    "google-cloud-storage>=3.10.1",
    "jupyter>=1.1.1",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
]
